Move blueprints and related functions
