+++ qcfractal/qcfractal/components/torsiondrive/routes.py
@@ -2,9 +2,9 @@ from typing import List
 
 from flask import current_app, g
 
-from qcfractal.api_v1.blueprint import api_v1
-from qcfractal.api_v1.helpers import wrap_route
 from qcfractal.flask_app import storage_socket
+from qcfractal.flask_app.api_v1.blueprint import api_v1
+from qcfractal.flask_app.api_v1.helpers import wrap_route
 from qcportal.exceptions import LimitExceededError
 from qcportal.torsiondrive import (
     TorsiondriveDatasetSpecification,
