+++ qcfractal/qcfractal/flask_app/api_v1/routes.py
@@ -1,6 +1,6 @@
 from flask import jsonify
 
-from qcfractal.api_v1.blueprint import api_v1
+from qcfractal.flask_app.api_v1.blueprint import api_v1
 
 
 @api_v1.route("/ping", methods=["GET"])
