+++ qcfractal/qcfractal/flask_app/auth_v1/routes.py
@@ -8,8 +8,8 @@ from flask_jwt_extended import (
     unset_jwt_cookies,
 )
 
-from qcfractal.auth_v1.blueprint import auth_v1
 from qcfractal.flask_app import storage_socket
+from qcfractal.flask_app.auth_v1.blueprint import auth_v1
 from qcfractal.flask_app.helpers import get_all_endpoints, access_token_from_user, login_and_get_jwt
 from qcportal.auth import UserInfo, RoleInfo
