+++ qcfractal/qcfractal/flask_app/handlers.py
@@ -16,7 +16,7 @@ from qcfractal.flask_app import storage_socket
 from qcfractal.flask_app.helpers import access_token_from_user
 from qcportal.auth import UserInfo, RoleInfo
 from qcportal.exceptions import UserReportableError, AuthenticationFailure, ComputeManagerError
-from .home import home_blueprint
+from .home_v1 import home_v1
 
 
 #####################################################################
@@ -27,7 +27,7 @@ from .home import home_blueprint
 #####################################################################
 
 
-@home_blueprint.before_app_request
+@home_v1.before_app_request
 def before_request_func():
     # Store timing information in the request/app context
     # g here refers to flask.g
@@ -39,7 +39,7 @@ def before_request_func():
         g.request_bytes = 0
 
 
-@home_blueprint.after_app_request
+@home_v1.after_app_request
 def after_request_func(response: Response):
     #################################################################
     # NOTE: Do not touch response.response! It may mess up streaming
@@ -116,7 +116,7 @@ def after_request_func(response: Response):
     return response
 
 
-@home_blueprint.app_errorhandler(InternalServerError)
+@home_v1.app_errorhandler(InternalServerError)
 def handle_internal_error(error):
     # For otherwise unhandled errors
     # Do not report the details to the user. Instead, log it,
@@ -149,26 +149,26 @@ def handle_internal_error(error):
         return jsonify(msg=tb), error.code
 
 
-@home_blueprint.app_errorhandler(HTTPException)
+@home_v1.app_errorhandler(HTTPException)
 def handle_http_exception(error):
     # This handles many errors, such as NotFound, Unauthorized, etc
     # These are all reportable to the user
     return jsonify(msg=str(error)), error.code
 
 
-@home_blueprint.app_errorhandler(UserReportableError)
+@home_v1.app_errorhandler(UserReportableError)
 def handle_userreport_error(error):
     # This handles any errors that are reportable to the user
     return jsonify(msg=str(error)), 400
 
 
-@home_blueprint.app_errorhandler(AuthenticationFailure)
+@home_v1.app_errorhandler(AuthenticationFailure)
 def handle_auth_error(error):
     # This handles Authentication errors (invalid user, password, etc)
     return jsonify(msg=str(error)), 401
 
 
-@home_blueprint.app_errorhandler(ComputeManagerError)
+@home_v1.app_errorhandler(ComputeManagerError)
 def handle_compute_manager_error(error: ComputeManagerError):
     # Handle compute manager errors
     return jsonify(msg=str(error)), 400
